% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FieldPrep-package.R
\docType{package}
\name{FieldPrep}
\alias{FieldPrep-package}
\alias{FieldPrep}
\title{Misc tools for prepping for field season}
\description{
A collection of functions for prepping for field season.
}
\author{
\strong{Maintainer}: David Hope \email{david.hope@ec.gc.ca} (\href{https://orcid.org/0000-0002-2140-4261}{ORCID})

Other contributors:
\itemize{
  \item Government of Canada [copyright holder, funder]
}

}
