% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hexes.R
\name{plot_hex}
\alias{plot_hex}
\title{Create a plot map for helicopter decisions}
\usage{
plot_hex(
  plot_id,
  plot_col_label,
  site_col_label,
  plot_sf_df,
  sites_sf_sf,
  near_water_course,
  water_body,
  crs_to_use = 3161
)
}
\arguments{
\item{plot_id}{string plot}

\item{crs_to_use}{}
}
\description{
Create a plot map for helicopter decisions
}
